+++ think/__init__.py
@@ -1,5 +1,5 @@
 from .cluster import cluster as cluster_day
-from .cluster import cluster_range
+from .cluster import cluster_range, cluster_scan
 from .detect_border import detect_border
 from .detect_created import detect_created
 from .detect_transcript import detect_transcript_json, detect_transcript_segment
@@ -9,6 +9,7 @@ from .sunstone import main as sunstone_main
 __all__ = [
     "cluster_day",
     "cluster_range",
+    "cluster_scan",
     "detect_border",
     "sunstone_main",
     "detect_created",